+++ app/Http/Livewire/Administration/AreaPossition/PossitionComponent.php
@@ -1,6 +1,7 @@
 <?php
 
 namespace App\Http\Livewire\Administration\AreaPossition;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use App\Models\Possition;
 use App\Models\Area;
 use Livewire\Component;
@@ -65,27 +66,42 @@ class PossitionComponent extends Component
     }
     public function update()
     {
-        $possition = Possition::findOrFail($this->codigo);        
-        $possition->update([
-            'idareas'=>$this->area,
-            'description'=>$this->descripcion,
-            'addnote'=>$this->notas,
-            'level'=>$this->nivel,
-        ]);
-        $this->limpiar();
-        $this->dispatchBrowserEvent('swal',[
-            'title'=>'Actualizado!',
-            'text'=>'La información se actualizó correctamente!',
-            'timer'=>3000,
-            'icon'=>'success',
-            'toast'=>true,
-            'position'=>'top-right'
-        ]);
+        try {
+            $this->authorize('Editar Cargo');
+            $possition = Possition::findOrFail($this->codigo);        
+            $possition->update([
+                'idareas'=>$this->area,
+                'description'=>$this->descripcion,
+                'addnote'=>$this->notas,
+                'level'=>$this->nivel,
+            ]);
+            $this->limpiar();
+            $this->dispatchBrowserEvent('swal',[
+                'title'=>'Actualizado!',
+                'text'=>'La información se actualizó correctamente!',
+                'timer'=>3000,
+                'icon'=>'success',
+                'toast'=>true,
+                'position'=>'top-right'
+            ]);
+        } catch (\Throwable $th) {
+            $this->limpiar();
+            $this->dispatchBrowserEvent('swal',[
+                'title'=>'No Actualizado!',
+                'text'=>'No se pudo actualizar',
+                'timer'=>3000,
+                'icon'=>'error',
+                'toast'=>true,
+                'position'=>'top-right'
+            ]);
+        }
+        
     }
 
     public function store()
     {
         try {
+            $this->authorize('Agregar Cargo');
             Possition::create([
                 'idareas'=>$this->area,
                 'description'=>$this->descripcion,
@@ -102,6 +118,7 @@ class PossitionComponent extends Component
                 'position'=>'top-right'
             ]);
         } catch (\Throwable $th) {
+            $this->limpiar();
             $this->dispatchBrowserEvent('swal',[
                 'title'=>'No Agregado!',
                 'text'=>'No se pudo agregar el nuevo registro',
@@ -118,6 +135,7 @@ class PossitionComponent extends Component
     public function destroy(){
 
         try {
+            $this->authorize('Eliminar Cargo');
             Possition::destroy($this->codigo);
             $this->limpiar();
             $this->dispatchBrowserEvent('swal',[
