+++ app/Http/Livewire/Maintenance/ChannelComponent.php
@@ -1,7 +1,7 @@
 <?php
 
 namespace App\Http\Livewire\Maintenance;
-
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use App\Models\Channel;
 use Livewire\Component;
 use Livewire\WithPagination;
@@ -55,40 +55,69 @@ class ChannelComponent extends Component
 
     public function update()
     {
-        $channel = Channel::findOrFail($this->codigo);        
-        $channel->update([
-            'description'=>$this->description
-        ]);
-        $this->limpiar();
-        $this->dispatchBrowserEvent('swal',[
-            'title'=>'Actualizado!',
-            'text'=>'La información se actualizó correctamente!',
-            'timer'=>3000,
-            'icon'=>'success',
-            'toast'=>true,
-            'position'=>'top-right'
-        ]);
+        try {
+            $this->authorize('Agregar Canales de Atención');
+            $channel = Channel::findOrFail($this->codigo);        
+            $channel->update([
+                'description'=>$this->description
+            ]);
+            $this->limpiar();
+            $this->dispatchBrowserEvent('swal',[
+                'title'=>'Actualizado!',
+                'text'=>'La información se actualizó correctamente!',
+                'timer'=>3000,
+                'icon'=>'success',
+                'toast'=>true,
+                'position'=>'top-right'
+            ]);
+        } catch (\Throwable $th) {
+            $this->limpiar();
+            $this->dispatchBrowserEvent('swal',[
+                'title'=>'No Actualizado!',
+                'text'=>'No se pudo actualizar',
+                'timer'=>3000,
+                'icon'=>'error',
+                'toast'=>true,
+                'position'=>'top-right'
+            ]);
+        }
+        
     }
 
     public function store()
     {
-        Channel::create([
-            'description'=>$this->description
-        ]);        
-        $this->limpiar();
-        $this->dispatchBrowserEvent('swal',[
-            'title'=>'Agregado!',
-            'text'=>'La información se agregó correctamente!',
-            'timer'=>3000,
-            'icon'=>'success',
-            'toast'=>true,
-            'position'=>'top-right'
-        ]);
+        try {
+            $this->authorize('Editar Canales de Atención');
+            Channel::create([
+                'description'=>$this->description
+            ]);        
+            $this->limpiar();
+            $this->dispatchBrowserEvent('swal',[
+                'title'=>'Agregado!',
+                'text'=>'La información se agregó correctamente!',
+                'timer'=>3000,
+                'icon'=>'success',
+                'toast'=>true,
+                'position'=>'top-right'
+            ]);
+        } catch (\Throwable $th) {
+            $this->limpiar();
+            $this->dispatchBrowserEvent('swal',[
+                'title'=>'No Agregado!',
+                'text'=>'No se pudo agregar el nuevo registro',
+                'timer'=>3000,
+                'icon'=>'error',
+                'toast'=>true,
+                'position'=>'top-right'
+            ]);
+        }
+        
     }
 
     public function destroy(){
 
         try {
+            $this->authorize('Eliminar Canales de Atención');
             Channel::destroy($this->codigo);
             $this->limpiar();
             $this->dispatchBrowserEvent('swal',[
